Use a library-wide instance of `IdFactory`.
This should guarantee that no two "persistent IDs" are ever
duplicated.
